+++ src/Sitecore.Pathfinder.Server/Emitters/EmitService.cs
@@ -1,11 +1,14 @@
 ﻿// © 2015 Sitecore Corporation A/S. All rights reserved.
 
+using System.IO;
+using Sitecore.Pathfinder.Configuration;
+
 namespace Sitecore.Pathfinder.Emitters
 {
-
     public enum EmitSource
     {
         NugetPackage,
+
         Directory
     }
 
@@ -17,20 +20,25 @@ namespace Sitecore.Pathfinder.Emitters
             EmitSource = emitSource;
         }
 
+        public EmitSource EmitSource { get; }
+
         [NotNull]
         public string ProjectDirectory { get; }
-
-        public EmitSource EmitSource { get; }
 
         public virtual void Start()
         {
             var startup = new Startup();
 
-            var configuration = startup.RegisterConfiguration(ProjectDirectory, EmitSource);
+            var configuration = ConfigurationStartup.RegisterConfiguration(ProjectDirectory, ConfigurationOptions.Noninteractive);
+            if (configuration == null)
+            {
+                return;
+            }
+
             var compositionService = startup.RegisterCompositionService(configuration);
 
             var emitter = compositionService.GetExportedValue<Emitter>();
-            emitter.Start();                    
+            emitter.Start();
         }
     }
 }